# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MDIA API",
        "description": "API documentation for Mbaise Diaspora in the Americas application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/reports/member": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "View member payment report",
                "description": "Retrieve full payment history for the authenticated member structured for printing with optional year filter",
                "operationId": "6df887a3b51cc13ab2489f0631890a9b",
                "parameters": [
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Filter payments by year",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2024
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member report retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Member report retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "member": {
                                                    "properties": {
                                                        "name": {
                                                            "type": "string",
                                                            "example": "John Doe"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "john@example.com"
                                                        },
                                                        "phone": {
                                                            "type": "string",
                                                            "example": "1234567890"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "payments": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "uuid": {
                                                                "type": "string",
                                                                "format": "uuid"
                                                            },
                                                            "payment_type": {
                                                                "type": "string",
                                                                "example": "Annual Membership"
                                                            },
                                                            "amount": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 100
                                                            },
                                                            "year": {
                                                                "type": "integer",
                                                                "example": 2024
                                                            },
                                                            "payment_date": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "2024-01-15"
                                                            },
                                                            "notes": {
                                                                "type": "string",
                                                                "nullable": true
                                                            },
                                                            "status": {
                                                                "type": "string",
                                                                "example": "verified"
                                                            },
                                                            "verified_by_name": {
                                                                "type": "string",
                                                                "example": "Admin User",
                                                                "nullable": true
                                                            },
                                                            "verified_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "nullable": true
                                                            },
                                                            "evidence_files": {
                                                                "type": "array",
                                                                "items": {
                                                                    "type": "string",
                                                                    "example": "/uploads/evidence/receipt.jpg"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}